requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
sqlparse==0.4.4
